nyay-ai/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic==2.9.0
python-multipart==0.0.12
python-dotenv==1.0.1
aiofiles==24.1.0
groq>=0.9.0
psycopg2-binary==2.9.9
